move sparkline component
