+++ frontend/src/scenes/error-tracking/OccurrenceSparkline.tsx
